# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: InstaVEG - Hyperlocal grocery marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wishlist/{productId}:
    post:
      operationId: addToWishlist
      tags: [users]
      summary: Add product to wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Added

    delete:
      operationId: removeFromWishlist
      tags: [users]
      summary: Remove product from wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed

  # USER PROFILE
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUserProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/addresses:
    get:
      operationId: listAddresses
      tags: [users]
      summary: List user addresses
      responses:
        "200":
          description: Addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
    post:
      operationId: addAddress
      tags: [users]
      summary: Add new address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "201":
          description: Address added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
  /users/addresses/{id}:
    patch:
      operationId: updateAddress
      tags: [users]
      summary: Update address
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "200":
          description: Updated address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
    delete:
      operationId: deleteAddress
      tags: [users]
      summary: Delete address
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # SELLERS
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [buyer, seller, admin]
        phone:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        isActive:
          type: boolean
        coins:
          type: integer
        createdAt:
          type: string
          format: date-time
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        avatar:
          type: string
...
    Address:
      type: object
      required: [id, label, street, city, state, pincode]
      properties:
        id:
          type: integer
        userId:
          type: integer
        label:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        isDefault:
          type: boolean
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
    AddressInput:
      type: object
      required: [label, street, city, state, pincode]
      properties:
        label:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        isDefault:
          type: boolean
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]

    # SELLER
...